Items were only received if the connection was up. This doesn't work for async at all, and prevents multiple users from using the same slot (if someone has to bail from a sync, for example).
Client.py:
Put the connection info into properties that are shared with the socket and with the currently running loop.
Socket.py:
Added an 'ack' command for the lua to send back, confirming that the item was received.
Confirmed items are stored in a local file, so that if a game file gets corrupted the local file can be deleted and the items all resend (or if someone needs to jump in from another location)

This might all be completely unnecessary with how Shananas has changed the lua